# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Renovix AI API
  description: Unified AI platform with multi-LLM API, chat portal, admin dashboard, RAG knowledge base, AI agents, and MCP integrations.
  version: 1.0.0
  contact:
    name: Renovix AI
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.renovix.id/api/v1
    description: Production
security: []
...
paths:
...
  /codes/redeem:
    post:
      tags: [Codes]
      summary: Redeem invitation code
      operationId: redeemCode
      security:
        - BearerAuth: []
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Code redeemed
          content:
            application/json:
              schema:
                type: object
                properties:
                  amount:
                    type: string
                  success:
                    type: boolean
  /codes/history:
    get:
      tags: [Codes]
      summary: Get user's code redemption history
      operationId: getCodeHistory
      security:
        - BearerAuth: []
        - SessionAuth: []
      responses:
        "200":
          description: Code history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodeHistory"

  # ─── Reminders ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
    SessionAuth:
      type: http
      scheme: bearer
      bearerFormat: session
      description: Better Auth session token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API key with Bearer prefix: `Bearer sk_live_*`"
    PortalSessionAuth:
      type: apiKey
      in: header
      name: X-Portal-Session
      description: Portal session token for anonymous/free-tier users
...
  schemas:
...
    CodeHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        amount:
          type: string
        redeemedAt:
          type: string
          format: date-time
...